add model for PageLink, convert PageArrow.js and PageLink.js to tsx
